Reorganize helper classes inside a separate package
